+++ rq/timeouts.py
@@ -33,7 +33,7 @@ class death_penalty_after(object):
 
     def handle_death_penalty(self, signum, frame):
         raise JobTimeoutException('Job exceeded maximum timeout '
-                'value (%d seconds).' % self._timeout)
+                                  'value (%d seconds).' % self._timeout)
 
     def setup_death_penalty(self):
         """Sets up an alarm signal and a signal handler that raises